Moved `version_check` to separate `compatibility_check.py` module as to not clutter `main.py` with arbitrary stuff.
